dev: add: complete refactoring
