+++ src/express/commands/start.js
@@ -1,10 +1,9 @@
-import {editMaticCliDockerYAMLConfig, editMaticCliRemoteYAMLConfig, splitToArray} from "../config-utils";
 import yaml from "js-yaml";
 import fs from "fs";
-import {runScpCommand, runSshCommand} from "../remote-worker";
-import {maxRetries} from "../constants";
-const shell = require("shelljs");
+import {editMaticCliDockerYAMLConfig, editMaticCliRemoteYAMLConfig, splitToArray} from "../common/config-utils";
+import {runScpCommand, runSshCommand, maxRetries} from "../common/remote-worker";
 
+const shell = require("shelljs");
 const timer = ms => new Promise(res => setTimeout(res, ms))
 
 async function terraformApply() {
