--- src/express/constants.js
@@ -1,7 +1,0 @@
-let remoteStdio = 'inherit'
-const maxRetries = 3
-
-module.exports = {
-    remoteStdio,
-    maxRetries,
-};
